requirements.txt
# ADM Platform - Vercel Serverless Dependencies
fastapi==0.109.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
anthropic==0.43.0
python-telegram-bot==21.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx>=0.24.0
gtts==2.5.1

# Database driver (PostgreSQL / Neon only - no SQLite on Vercel)
psycopg2-binary==2.9.9
